# api/openapi.yaml
openapi: 3.0.3
info:
  title: HomeIOT Device API
  version: 1.0.0
  description: |
    Tightened v1 contract for HomeIOT device-server integration.

    Rules:
    - Authentication headers are mandatory on every endpoint.
    - Header identity is authoritative (X-Device-ID).
    - Heartbeat returns only mode-based runtime control (no role/admin_exec_enabled).
    - Development mode is additive: modules still run, and dev command polling is enabled.
    - Dev commands are execute-on-change (revision_hash) unless forceRerun=true.
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - DeviceIdHeader: []
    ApiKeyHeader: []
paths:
  /api/devices/register:
    post:
      summary: Register or refresh device record
      operationId: registerDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered or refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '201':
          description: Newly registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid payload or body/header device identity mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/devices/heartbeat:
    post:
      summary: Send device heartbeat and receive runtime metadata
      operationId: heartbeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeartbeatRequest'
      responses:
        '200':
          description: Heartbeat accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeartbeatResponse'
        '400':
          description: Invalid payload or body/header mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/devices/logs:
    post:
      summary: Upload batched logs
      operationId: uploadLogs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogBatchRequest'
      responses:
        '200':
          description: Logs accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    DeviceIdHeader:
      type: apiKey
      in: header
      name: X-Device-ID
      description: Device identity header (authoritative).
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Device API key.
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    StatusResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]
    RegisterRequest:
      type: object
      properties:
        device_id:
          type: string
        platform:
          type: string
          enum: [esp32, pico]
        version:
          type: string
        ip:
          type: string
        timestamp:
          type: integer
      required: [device_id, version]
    RegisterResponse:
      type: object
      properties:
        status:
          type: string
        device_id:
          type: string
      required: [status, device_id]
    HeartbeatRequest:
      type: object
      properties:
        device_id:
          type: string
        timestamp:
          type: integer
        uptime_ms:
          type: integer
        free_memory_bytes:
          type: integer
      required: [device_id, timestamp]
    HeartbeatResponse:
      type: object
      properties:
        status:
          type: string
        server_time_utc:
          type: string
          format: date-time
        mode:
          type: string
          enum: [production, development]
        dev_poll_interval_ms:
          type: integer
        module_assignment_poll_interval_ms:
          type: integer
        next_heartbeat_ms:
          type: integer
      required: [status, mode]
...
    LogEntry:
      type: object
      properties:
        ts:
          type: integer
        level:
          type: string
        message:
          type: string
        context:
          type: object
          additionalProperties: true
      required: [ts, level, message]
    LogBatchRequest:
      type: object
      properties:
        device_id:
          type: string
        reason:
          type: string
        sentAt:
          type: integer
        dropped_count:
          type: integer
        truncated:
          type: boolean
        logs:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'
      required: [device_id, reason, sentAt, logs]
...